dynamic hdr change

+++ src/main/java/OxyEngine/OxyApplication.java
@@ -16,7 +16,7 @@ public abstract class OxyApplication implements OxyDisposable {
         layerStack = new LayerStack();
     }
 
-    public static int FPS = 0;
+    public static float FPS = 0;
 
     public abstract void init();
